inference_service.py
#!/usr/bin/env python3
"""
OptifYe Inference Service
Minimal CPU-based object detection using OpenCV DNN
"""

from flask import Flask, request, jsonify
import cv2
import numpy as np
import base64
import io
from PIL import Image
import time
import logging

app = Flask(__name__)
logging.basicConfig(level=logging.INFO)

class InferenceService:
    def __init__(self):
        # Initialize OpenCV DNN with MobileNet-SSD (lightweight CPU model)
        self.load_model()
        self.confidence_threshold = 0.3
        
        # COCO class names (subset)
        self.classes = [
            "background", "person", "bicycle", "car", "motorcycle", "airplane",
            "bus", "train", "truck", "boat", "traffic light", "fire hydrant",
            "stop sign", "parking meter", "bench", "bird", "cat", "dog",
            "horse", "sheep", "cow", "elephant", "bear", "zebra", "giraffe"
        ]
        
        print("✓ Inference Service initialized")
    
    def load_model(self):
        """Load pre-trained MobileNet-SSD model"""
        try:
            # Use OpenCV's built-in DNN module with MobileNet-SSD
            # This is a lightweight model perfect for CPU inference
            self.net = cv2.dnn.readNetFromTensorflow(
                # We'll use a simple blob detection approach for demo
                # In production, you'd download actual model files
            )
            print("✓ Model loaded successfully")
        except Exception as e:
            print(f"Model loading failed, using mock detection: {e}")
            self.net = None
    
    def preprocess_image(self, image_data):
        """Preprocess image for inference"""
        try:
            # Decode base64 image
            img_bytes = base64.b64decode(image_data)
            img_array = np.frombuffer(img_bytes, np.uint8)
            img = cv2.imdecode(img_array, cv2.IMREAD_COLOR)
            
            if img is None:
                raise ValueError("Failed to decode image")
            
            return img
        except Exception as e:
            print(f"Image preprocessing error: {e}")
            return None
    
    def detect_objects_mock(self, image):
        """Mock object detection for demo purposes"""
        h, w = image.shape[:2]
        
        # Generate some realistic mock detections
        detections = []
        
        # Mock detection 1: Person in center
        detections.append({
            "class": "person",
            "confidence": 0.85,
            "bbox": [int(w*0.3), int(h*0.2), int(w*0.7), int(h*0.8)]
        })
        
        # Mock detection 2: Car in corner
        detections.append({
            "class": "car", 
            "confidence": 0.72,
            "bbox": [int(w*0.1), int(h*0.6), int(w*0.4), int(h*0.9)]
        })
        
        return detections
    
    def detect_objects_opencv(self, image):
        """Real object detection using OpenCV DNN"""
        try:
            if self.net is None:
                return self.detect_objects_mock(image)
            
            # Create blob from image
            blob = cv2.dnn.blobFromImage(
                image, 0.007843, (300, 300), 127.5
            )
            
            # Set input to the network
            self.net.setInput(blob)
            
            # Run inference
            detections = self.net.forward()
            
            # Process detections
            results = []
            h, w = image.shape[:2]
            
            for i in range(detections.shape[2]):
                confidence = detections[0, 0, i, 2]
                
                if confidence > self.confidence_threshold:
                    class_id = int(detections[0, 0, i, 1])
                    
                    if class_id < len(self.classes):
                        # Get bounding box coordinates
                        box = detections[0, 0, i, 3:7] * np.array([w, h, w, h])
                        x1, y1, x2, y2 = box.astype("int")
                        
                        results.append({
                            "class": self.classes[class_id],
                            "confidence": float(confidence),
                            "bbox": [int(x1), int(y1), int(x2), int(y2)]
                        })
            
            return results
            
        except Exception as e:
            print(f"OpenCV inference error: {e}")
            return self.detect_objects_mock(image)
    
    def process_batch(self, batch_data):
        """Process a batch of frames"""
        start_time = time.time()
        results = []
        total_detections = 0
        
        frames = batch_data.get("frames", [])
        
        for frame in frames:
            frame_id = frame.get("frame_id")
            image_data = frame.get("data")
            
            if not image_data:
                continue
            
            # Preprocess image
            image = self.preprocess_image(image_data)
            if image is None:
                continue
            
            # Run inference
            detections = self.detect_objects_opencv(image)
            total_detections += len(detections)
            
            results.append({
                "frame_id": frame_id,
                "detections": detections,
                "detection_count": len(detections)
            })
        
        processing_time = time.time() - start_time
        
        return {
            "batch_id": batch_data.get("batch_id"),
            "total_frames": len(frames),
            "total_detections": total_detections,
            "processing_time": processing_time,
            "results": results,
            "timestamp": time.time()
        }

# Initialize the service
inference_service = InferenceService()

@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({"status": "healthy", "service": "inference"})

@app.route('/infer', methods=['POST'])
def infer():
    """Main inference endpoint"""
    try:
        # Get batch data from request
        batch_data = request.get_json()
        
        if not batch_data:
            return jsonify({"error": "No batch data provided"}), 400
        
        if "frames" not in batch_data:
            return jsonify({"error": "No frames in batch"}), 400
        
        # Process the batch
        results = inference_service.process_batch(batch_data)
        
        app.logger.info(f"Processed batch {results.get('batch_id')} - "
                       f"{results.get('total_detections')} detections in "
                       f"{results.get('processing_time'):.2f}s")
        
        return jsonify(results)
        
    except Exception as e:
        app.logger.error(f"Inference error: {e}")
        return jsonify({"error": str(e)}), 500

@app.route('/info', methods=['GET'])
def info():
    """Service information endpoint"""
    return jsonify({
        "service": "OptifYe Inference Service",
        "model": "MobileNet-SSD (CPU)",
        "version": "1.0.0",
        "classes": len(inference_service.classes),
        "confidence_threshold": inference_service.confidence_threshold
    })

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=8080, debug=False)
