requirements_inference.txt
Flask==2.3.3
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
